Add script to forward model Noddyverse slices
Noddyverse models are at 100 m altitude. My next work requires 3D potential fields to sample from. This requires remodelling Noddyverse models at 20 m sampling altitude slices.

I selected 100, 4000, 20, so as to have isotropic voxels - x,y, are 20 m, so z at 20 m makes sense, even if it is a little excessive.

It would be useful to rewrite this code to use ntasks with altitude chunks - thanks @LLgeo for the suggestion! But it's not crucial to the performance I require.